+++ aidial_adapter_openai/video_generation/azure/client.py
